// frontend/src/client/api/generated/model/transactionRead.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionReadSubcategory } from "./transactionReadSubcategory";
import type { EnumTransactionType } from "./enumTransactionType";
import type { TransactionReadMemo } from "./transactionReadMemo";
import type { TransactionReadOrganizationId } from "./transactionReadOrganizationId";

export interface TransactionRead {
  id: string;
  date: string;
  category: string;
  subcategory?: TransactionReadSubcategory;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  amount: string;
  type: EnumTransactionType;
  payment_method: string;
  description: string;
  memo?: TransactionReadMemo;
  hash: string;
  created_at: string;
  updated_at: string;
  organization_id?: TransactionReadOrganizationId;
}